% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pterm.R
\name{control.aic}
\alias{control.aic}
\title{Control function for aic, caic, or epic-based optimization of a
  \code{coxph.penalty} term}
\usage{
control.aic(parms, iter, old, n, df, loglik)
}
\description{
Control function for aic, caic, or epic-based optimization of a
  \code{coxph.penalty} term
}
\references{
Therneau, T. M. and Grambsch, P. M. (1998). Penalized Cox models and Frailty.
Technical report, Division of Biostatistics. Mayo Clinic; Rochester, MN,
pages 1–58. Available at
http://www.mayo.edu/research/documents/frailtypdf/doc-10027273.
}
\keyword{internal}

