% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.xt.func}
\alias{create.xt.func}
\title{Create x-transform function for non-time-varying terms}
\usage{
create.xt.func(limits, linear, basistype, sind, integration, standardize,
  domain, basisargs, method, eps)
}
\description{
Create x-transform function for non-time-varying terms
}
\keyword{internal}

