% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcox-Accessory.R
\name{expand.call}
\alias{expand.call}
\title{Expands a call}
\usage{
expand.call(definition = NULL, call = sys.call(sys.parent(1)),
  expand.dots = TRUE)
}
\description{
This function was copied from the \code{refund} package.
}
\keyword{internal}

