% Generated by roxygen2 (4.0.1): do not edit by hand
\name{f_override}
\alias{f_override}
\title{Accessory function that allows s(), te(), and t2() in coxph term names}
\usage{
f_override(...)
}
\description{
Accessory function that allows s(), te(), and t2() in coxph term names
}
\keyword{internal}

