% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genX}
\alias{genX}
\title{Simulate smooth time-varying (funcitonal) covariates}
\usage{
genX(N, s = NULL)
}
\arguments{
\item{N}{sample size (number of rows of \code{X})}

\item{s}{vector of time points at which the covariate is observed.
The length of \code{s}, \code{J}, is used for the number of columns
of \code{X}. Defaults to \code{0:100}.}
}
\value{
An \code{N x J} matrix \code{X}
}
\description{
This function simulates functional covariates, which could also be
considered to be a densely-sampled time-varying covariate.
The covariates are not truncated,
so the function will return a full grid of values that the covariate
could take if observed for that long.
}
\details{
This function generates a functional covariate \code{X}
from the following model:
}
\author{
Jonathan Gellar <jgellar1@jhu.edu>
}
\seealso{
\code{\link{simTVSurv}}
}

