% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_pcox.R
\name{geom_pcox1}
\alias{geom_pcox1}
\title{Plot univariate \code{pcox} coefficients}
\usage{
geom_pcox1(yvar, se = se, type = c("ribbon", "line"), line_opts = list(),
  ci_opts = list(), bw = TRUE)
}
\arguments{
\item{yvar}{the variable containing the estimate values}

\item{se}{the variable containing the standard errors. Set to \code{NULL}
to omit the confidence interval}

\item{type}{type of confidence interval, either a ribbon or line}

\item{line_opts}{list of additional options for drawing the estimate, to be
passed to \code{geom_line}}

\item{ci_opts}{list of additional options for drawing the confidence interval,
to be passed to \code{geom_ribbon} or \code{geom_line}}

\item{bw}{if \code{TRUE}, uses \code{\link[ggplot2]{theme_bw()}}}
}
\description{
Plot univariate \code{pcox} coefficients
}
\examples{
# Put examples here
}

