% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_pcox.R
\name{geom_pcox2}
\alias{geom_pcox2}
\title{Plot bivariate \code{pcox} coefficients as a heatmap using \code{ggplot2}}
\usage{
geom_pcox2(z, colours = rev(RColorBrewer::brewer.pal(11, "Spectral")),
  zlim = NULL, expand.x = c(0, 0), expand.y = c(0, 0), bw = TRUE, ...)
}
\arguments{
\item{z}{variable to be represented by the color of the heatmap}

\item{colours}{color palette to use for both the \code{fill} and \code{colour}
aesthetics}

\item{expand.x}{amount to expand grid in x direction}

\item{expand.y}{amount to expand grid in y direction}

\item{bw}{if \code{TRUE}, uses \code{theme_bw()}}
}
\description{
This \code{geom} calls \code{\link[ggplot2]{geom_tile}}, and modifies some of
  the plot parameters to produce pretty heatmaps of bivariate \code{pcox}
  coefficients.
}

