% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.termname}
\alias{get.termname}
\title{Accessory function to extract the term name from a tt or xt function}
\usage{
get.termname(func, varnames = NULL)
}
\description{
Accessory function to extract the term name from a tt or xt function
}
\keyword{internal}

