% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcox-Accessory.R
\name{getCall.pcox}
\alias{getCall.pcox}
\title{\code{getCall} method for \code{pcox} class}
\usage{
\method{getCall}{pcox}(x)
}
\description{
\code{getCall} method for \code{pcox} class
}
\keyword{internal}

