% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getHCEst.R
\name{getHCEst2}
\alias{getHCEst2}
\title{Extract a historical Cox estimate from a fitted model}
\usage{
getHCEst2(fit, term = NULL, idx = NULL, mask = NULL)
}
\description{
Extract a historical Cox estimate from a fitted model
}

