% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getL.R
\name{getL2}
\alias{getL2}
\title{Get the weight matrix for a linear functional term}
\usage{
getL2(tind, integration, n.int = NULL)
}
\description{
Get the weight matrix for a linear functional term
}
\keyword{internal}

