% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getL.R
\name{getL3}
\alias{getL3}
\title{Get the weight matrix for a linear functional term}
\usage{
getL3(tind, integration, mask = NULL)
}
\description{
Get the weight matrix for a linear functional term
}
\keyword{internal}

