% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getROC.R
\name{getROC}
\alias{getROC}
\title{Generate Time-Varying ROC Curve}
\usage{
getROC(fit, evaltimes = NULL, plot = FALSE, K = NULL, data = NULL,
  bhaz = c("none", "breslow"))
}
\description{
Generate Time-Varying ROC Curve
}

