% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hlf}
\alias{hlf}
\title{Constuct a Cox regression term as a historical linear function}
\usage{
hlf(X, tind = NULL, limits = NULL, ...)
}
\arguments{
\item{X}{matrix containing the time-varying covariate. Should be
\eqn{N x J}, where \eqn{N} is the number of subjects and \eqn{J}
is the maximum number of time points per subject. Most rows will
have \code{NA} values in the right-most columns, corresponding
to unobserved time points.}

\item{tind}{matrix containing the time indices of evaluations of
   \eqn{X_i(t)}. If a matrix, it must be the same dimensionality
   as \code{X}; if a vector, must be of length \code{ncol(X)} and
   the same grid is assumed for all subjects; if omitted, the same
   equally-spaced grid is assumed for all subjects.}
}
\value{
A list with some stuff
}
\description{
This function defines a Cox regression term as a historical linear function
for inclusion in a \code{pcox} formula. This term treats a densely measured
time-varying covariate (i.e., biomarker) as a functional predictor,
and models its effect on the hazard at time t with as a smooth
funcitonal term. The default is the time-static term
\eqn{1\t \int_0^t X_i(s)\beta(s)ds}. A common modification would be
to make the term time-varying by specifying it as a\code{\link{tv}}
term, resulting in the formula \eqn{1\t \int_0^t X_i(s)\beta(s,t)ds}.
The \eqn{1/t} part of the formula may be removed and the integration
limits can be changed.
}
\details{
These are the details of the function
}
\author{
Jonathan E. Gellar <jgellar1@jhu.edu>
}

