% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcox-Accessory.R
\name{na.omit_pcox}
\alias{na.omit_pcox}
\title{Handle Missing Values for \code{pcox} Fits}
\usage{
na.omit_pcox(object, ...)
}
\description{
This function differs from \code{na.omit} in the way it handles data frames
that contain matrix elements. \code{na.omit} will remove an observation
if any of the elements in that row of the matrix are missing, whereas this
function only removes the observation if the entire row is missing.
}
\keyword{internal}

