% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcox.R
\name{pcox}
\alias{pcox}
\title{Fit a Penalized Cox Regression Model}
\usage{
pcox(formula, data, method = c("aic", "caic", "epic"), eps = 1e-06,
  knots = NULL, x = TRUE, ...)
}
\arguments{
\item{formula}{a formula, with special terms indicating penalized covariates;
see details.}

\item{data}{an optional data frame containing the variables in the model. If
not found in \code{data}, the variables are taken from
\code{environment(formula)}, typcially he environment from which
\code{pcox} is called.}

\item{method}{method used to optimize the smoothing parameter(s). \code{"aic"}
is the Akaike Informaion Criterion, \code{"caic"} is the corrected AIC
of Hurvich, et al (1998), and \code{"epic"} is the EPIC criterion of
Shinohara, et al (2011).}

\item{eps}{convergence level for the criterion indicated by \code{method}.}

\item{knots}{optional list containing the user-specified knot values for
basis construction, as in \code{\link[mgcv]{gam}}.}

\item{x}{should the design matrix be included in the return object?
Defaults to \code{TRUE}.}

\item{...}{additional arguments for \code{\link[survival]{coxph}} or
\code{\link[coxme]{coxme}}.}
}
\value{
a fitted \code{pcox} object. This is a \code{coxph} object with
additional information in the \code{pcox} entry.
}
\description{
This function fits a penalized Cox regression model using penalized splines 
to model the smooth coefficients. It is a wrapper for \code{survival::coxph}
to fit Cox models with smooth covariate effects for
(possibly censored) time-to-event data, of the general form \eqn{log[h_i(t)]
= log[h_0(t)] + f_1(z_i, t) + f_2(X_i(s), t)} where \eqn{h_i(t)} is the
hazard function for subject i, \eqn{h_0(t)} is the baseline hazard function,
\eqn{z_i} are scalar covariates, and \eqn{X_i(s)} are functional covariates.
Penalized covariates are set up using functions from the \code{mgcv} package
to define basis functions and penalization.
}
\details{
Penalized covariates are specified as special terms in the model formula.
 The following terms are allowed:
 \enumerate{
 \item{\code{\link{p}()}: Specifies a generic penalized term. This is primarily
   used for smooth functions of scalar covariates (e.g., \eqn{f(x_i)}) or
   time-varying effects of scalar covariates (\eqn{\beta(t) x_i})}. However,
   all penalized terms in \code{pcox} formula may be specified using \code{p}
   with the appropriate \code{limits} argument. In fact, the other special
   terms below are simply wrappers for \code{p}.
 \item{\code{\link{cf}()}: Wrapper for \code{p} to specify a concurrent or
   lagged effect of a time-varying covariate (concurrent functional
   predictor), e.g. \eqn{\beta x_i(t)} or \eqn{\beta x_i(t-\delta_0)}.}
 \item{\code{\link{bf}()}: Wrapper for \code{p} to specify a baseline
   functional predictor, e.g. \eqn{\int X_i(s)\beta(s) ds}.}
 \item{\code{\link{hf}()}: Wrapper for \code{p} to specify a historical
   functional predictor, e.g. \eqn{\int_0^t X_i(s)\beta(s,t) ds}.}
 }
 See the help files for each of the above terms for details regarding usage.
 In addition to these special terms, \code{\link[survival]{frailty}},
 \code{\link[survival]{strata}}, \code{\link[survival]{cluster}}, and \code{tt} terms are
 allowed, as in a \code{\link[survival]{coxph}} formula.
}
\examples{
# Generate some data
N <- 500
J <- 200
x <- runif(N, 0, 2*pi)
male <- rbinom(N, size = 1, prob=.5)

# Simple linear terms (can be done with coxph)
eta1 <- matrix(.6*x + .75*male, nrow=N, ncol=J)
dat1 <- simTVSurv(eta1, data.frame(x=x, male=male))
fit1 <- pcox(Surv(time, event) ~ x + male, data=dat1)
fit1b<- survival::coxph(survival::Surv(time,event) ~ x + male, data=dat1)
all.equal(fit1$coefficients, fit1b$coefficients)

# Smooth function of scalar x
eta2 <- matrix(sin(x) + .75*male, nrow=N, ncol=J)
dat2 <- simTVSurv(eta2, data.frame(x=x, male=male))
fit2 <- pcox(Surv(time, event) ~ p(x, linear=FALSE) + male, data=dat2)
est2 <- coef(fit2)
#plot(value ~ x, data=est2, type="l", lwd=1.5, ylim=1.25*range(est2$value))
#lines(sin(seq(0,2*pi,by=.1)) ~ seq(0,2*pi,by=.1), col="red")
#lines(value + 1.96*se ~ x, data=est2, lty=2)
#lines(value - 1.96*se ~ x, data=est2, lty=2)
#legend("topright", c("Estimate", "CI", "Truth"),
#       lty=c(1,2,1), col=c(1,1,2))

# See help files for p, cf, bf, and hf for more examples
}
\author{
Jonathan Gellar <JGellar@mathematica-mpr.com> and Fabian Scheipl
}
\references{
Gellar, Jonathan E., Elizabeth Colantuoni, Dale M. Needham, and Ciprian M.
Crainiceanu (2015). Cox Regression Models with Functional Covariates for Survival
Data. Statistical Modeling, 15 (3): 256-278.

Hurvich, C. M., Simonoff, J. S., and Tsai, C. L. (1998). Smoothing Parameter
Selection in Nonparametric Regression Using an Improved Akaike Information
Criterion. Journal of the Royal Statistical Society. Series B
(Methodological): 60(2), 271-293.

Shinohara, R., Crainiceanu, C., Caffo, B., and Reich, D. (2011). Longitudinal
Analysis of Spatiotemporal Processes : A Case Study of Dynamic Contrast-Enhanced
Magnetic Resonance Imaging in Multiple Sclerosis. Retrieved from
http://biostats.bepress.com/jhubiostat/paper231/.
}
\seealso{
\code{mgcv}'s \code{\link[mgcv]{smooth.terms}} for details of 
  \code{mgcv} syntax and available spline bases and penalties; the related 
  \code{\link[refund]{pfr}} and \code{\link[refund]{pffr}} from 
  \code{refund}.
}

