% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{pcoxp}
\alias{pcox-package}
\alias{pcoxp}
\alias{pcoxp-package}
\title{Penalized Cox regression models}
\description{
\code{\link{pcox}()} fits penalized Cox regression models using penalized splines
to model the smooth coefficients.
}

