% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pcoxTerm}
\alias{pcoxTerm}
\title{Create a smooth and coxph.penalty object for use in a pcox formula}
\usage{
pcoxTerm(data, limits, linear, tv, basistype, sind, integration, standardize,
  domain, basisargs, method, eps, smooth, t = NULL)
}
\description{
Create a smooth and coxph.penalty object for use in a pcox formula
}

