% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smooth.construct.pi.smooth.spec.R
\name{Predict.matrix.pi.smooth}
\alias{Predict.matrix.pi.smooth}
\title{Predict.matrix method for pi basis}
\usage{
\method{Predict.matrix}{pi.smooth}(object, data)
}
\description{
Predict.matrix method for pi basis
}

