% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.pcox.R
\name{predict.pcox}
\alias{predict.pcox}
\title{Predictions for a \code{pcox} model}
\usage{
\method{predict}{pcox}(object, type = c("lp", "risk", "terms", "lpmatrix"),
  newdata = NULL, indices = NULL, ptimes = NULL, stimes = NULL,
  n = NULL, se.fit = c("none", "var", "var2"), ...)
}
\arguments{
\item{object}{Object of class \code{"pcox"}}

\item{type}{type of prediction. Choices are the linear predictor
(\code{"lp"}), the risk score exp(lp) (\code{"risk"}), the
terms of the linear predictor (\code{"terms"})}

\item{newdata}{new data}

\item{indices}{the indices}

\item{ptimes}{the prediction times}

\item{stimes}{survival times for input data}

\item{n}{number of observations in new data}

\item{se.fit}{A flag indicating if (model-based) standard errors are to be
returned (NOT IMPLEMENTED YET)}

\item{...}{extra arguments (currently unused)}
}
\value{
MISSING
}
\description{
This function calculates predictions for a pcox model
}
\details{
These are the MISSING details
}
\author{
Jonathan Gellar <jgellar1@jhu.edu>
}
\seealso{
\code{\link{pcox}}
}

