% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pterm.R
\name{pterm}
\alias{pterm}
\title{Create a \code{coxph.penalty} object from an \code{mgcv}-style smooth
object}
\usage{
pterm(sm, method = c("aic", "caic", "epic", "df", "fixed"), eps = 1e-06)
}
\arguments{
\item{sm}{the \code{mgcv}-style smooth object}

\item{method}{method for optimizing the smoothing parameter. Note that only
\code{aic}, \code{caic}, and \code{epic} have been tested fully}

\item{eps}{tolerance level for optimizing the smoothing parameter}
}
\value{
a \code{coxph.penalty} object, which is a design matrix with a
  number of atrributes that define the penalty. See Therneau and Grambsch
  (1988) for details.
}
\description{
This function takes a an \code{mgcv}-style smooth object, created using
\code{\link[mgcv]{smoothCon}} or \code{\link[mgcv]{smooth.construct}},
and converts it into a \code{coxph.penalty} object that can be used
in a \code{survival::coxph} formula. It serves as the primary bridge
between the \code{mgcv} and \code{survival} packages.
}
\author{
Jonathan Gellar <jgellar1@jhu.edu>
}
\references{
Therneau, T. M. and Grambsch, P. M. (1998). Penalized Cox models and Frailty.
Technical report, Division of Biostatistics. Mayo Clinic; Rochester, MN,
pages 1–58. Available at
http://www.mayo.edu/research/documents/frailtypdf/doc-10027273.
}
\seealso{
\code{\link[survival]{coxph}} for a brief description of
  \code{coxph.penalty} terms; \code{\link[mgcv]{smoothCon}} and
  \code{\link[mgcv]{smooth.construct}} for describing \code{mgcv}-style
  smooth objects; \code{\link[survival]{pspline}},
  \code{\link[survival]{ridge}}, or \code{\link[survival]{frailty}} for
  other examples of \code{coxph.penalty} terms
}
\keyword{internal}

