% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/risksetROC2.R
\name{risksetROC2}
\alias{risksetROC2}
\title{risksetROC 2nd version}
\usage{
risksetROC2(Stime, entry = NULL, status, marker, predict.time, span = NULL,
  order = 1, window = "asymmetric", prop = 0.5, plot = TRUE,
  type = "l", xlab = "FP", ylab = "TP", ...)
}
\description{
risksetROC 2nd version
}
\keyword{internal}

