% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcox_Utilities.R
\name{safeDeparse}
\alias{safeDeparse}
\title{Deparse a function safely}
\usage{
safeDeparse(expr)
}
\arguments{
\item{An}{expression

This utility function was copied from the \code{refund} package.}
}
\description{
Deparse a function safely
}

