% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simTVSurv}
\alias{simTVSurv}
\title{Simulate time-varying survival data}
\usage{
simTVSurv(etaMat, Xdat, timeIndex = NULL, eventRandom = NULL,
  censorRandom = NULL, groupByD = FALSE)
}
\arguments{
\item{etaMat}{an \code{N x J} matrix contining the linear predictor
for each subject at each time index, where \code{N} is the number
of subjects and \code{J} is the number of time points.}

\item{Xdat}{an \code{N x p} data frame or matrix, containing the
covariates used in the model, where \code{p} is the number of
covariates. Time-varying covariates are included as an
\code{N x J} matrix using the \code{\link{I}} ("AsIs") function;
see Details.}

\item{timeIndex}{A vector of length \code{J} containing the time
indices corresponding to the columns of \code{etaMat}; defaults
to \code{1:J}.}

\item{eventRandom}{individual event times. Can either be a vector
non-negative integer values, or a random generating function
with argument \code{n}. See Details.}

\item{censorRandom}{individual censoring times, similar to
\code{eventRandom}.}

\item{groupByD}{Logical, should individual assignment of event and
censoring times be replaced with group assignments? Defaults to
\code{FALSE}. See Details.}
}
\value{
A data frame containing the simulated data, as well as
any predictors supplied in \code{Xdat}
}
\description{
This is a modification of the \code{PermAlgo::permalgorithm()}
function, which implements the permutation algorithm of Sylvestre
and Abrahamowicz to simulate survival data with time-varying
coefficients. This version simulates time-varying survival data
under more general conditions. In particular, the function can
simulate data with time-varying coefficients, time-varying effects,
or combinations of the two.
}
\details{
We assume the following Cox-type regression model:

\eqn{log(h_i(t)) = log(h_0(t)) + \eta_i(t)}

where \eqn{h_i(t)} is the hazard function for subject \eqn{i}
at time \eqn{t}, \eqn{h_0(t)} is the baseline hazard function,
and \eqn{\eta_i(t)} is the linear predictor matrix, which is
a function of the covariates. As opposed to
\code{permalgorithm}, which \eqn{\eta_i(t)} by having the
user supply a vector of \eqn{\beta} coefficients and assuming
a concurrent, non-time-varying effect for time-varying
covariates, \code{simTVSurv} accepts \eqn{\eta_i(t)} directly
as an argument. Though this approach requires the user to do
some more calculations outside of this function, it allows added
flexibility by letting the user create the matrix using whatever
model they prefer. In particular, this allows the user to
specify time-varying effects for both time-fixed and time-varying
covariates, and non-concurrent effects of time-varying covariates.

\code{Xdat} is only required as an argument so that the covariates
will be included in the output dataset.
Any time-varying predictors must first be organized into an
\code{N x J} matrix, indicating a value for the covariate for
each subject at each \code{timeIndex}. The matrix is then included
as a "column" of \code{Xdat} (which must be a data frame), by
specifying it as an "AsIs" term with \code{I()}. The dataset that
is outputed by \code{simTVSurv} will modify the matrices for any
time-varying predictors so that any time points after an event or
censoring with be replaced with \code{NA}.

\code{eventRandom} and \code{censorRandom} can each be specified as
either a vector of non-negative integer values, or as a random
generating function with argument \code{n}. In either case, the
values must be smaller or equal to \code{J}. Both
default to Uniform[1,\code{J}].
\code{groupByD} is an option that, when enabled, increases the
compuational efficiency of the algorithm by replaceing the
individual assignment of event times and censoring times by
grouped assignments. The side effect of this option is that it
generates datasets that are, on average, slightly less consistent
with the model described by \code{etaMat} than those generated when
\code{groupByD=FALSE}. Still, \code{groupByD=TRUE} may be useful
to generate large datasets where \code{J<<N} so that many ties
are expected.

This is my link to \code{\link{pcox}}.
}
\author{
Jonathan Gellar <jgellar1@jhu.edu>
}
\seealso{
\code{\link{permalgorithm}}
}

