% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survfit.pcox.R
\name{survfit.pcox}
\alias{survfit.pcox}
\title{Compute a Survival Curve from a \code{pcox} model}
\usage{
\method{survfit}{pcox}(formula, newdata, se.fit = TRUE, conf.int = 0.95,
  individual = FALSE, type, vartype, conf.type = c("log", "log-log",
  "plain", "none"), censor = TRUE, id, na.action = na.pass, ...)
}
\arguments{
\item{formula}{a \code{pcox} object}
}
\value{
An object of class \code{"survfit"}. See \code{\link[survival]{survfit.object}}
  for details. Methods defined for \code{survfit} objects are \code{print},
  \code{plot}, \code{lines}, and \code{points}.
}
\description{
Computes the predicted survivor function for a \code{pcox} model. This
function is a modified version of \code{\link[survival]{survfit.coxph}}.
}
\details{
Any details needed?
}
\author{
Jonathan Gellar <JGellar@mathematica-mpr.com>
}
\seealso{
\code{\link[survival]{survfit}}, \code{\link[survival]{survfit.coxph}},
  \code{\link{pcox}}
}

