% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coefficients.pcox.R
\name{coefficients.pcox}
\alias{coef.pcox}
\alias{coefficients.pcox}
\title{Extract coefficient functions from a fitted pcox-object}
\usage{
\method{coefficients}{pcox}(object, select = NULL, se = TRUE,
  use.var2 = FALSE, n = NULL, n2 = NULL, exclude = FALSE,
  limit = TRUE, plotMe = FALSE, ...)

\method{coef}{pcox}(object, select = NULL, se = TRUE, use.var2 = FALSE,
  n = NULL, n2 = NULL, exclude = FALSE, limit = TRUE, plotMe = FALSE,
  ...)
}
\arguments{
\item{object}{return object from \code{\link{pcox}}}

\item{select}{integer vector indicating the index of the desired smooth term
in \code{object$pcox$smooth}. If \code{NULL}, returns all smooth terms.}

\item{se}{if \code{TRUE}, returns pointwise standard error estimates. If
numeric, the standard error will be multiplied by that number (e.g., 2
to make it easier to plot confidence intervals).}

\item{use.var2}{if \code{TRUE}, uses \code{object$var2} as the covariance
matrix for standard error estimation; otherwise uses \code{object$var}.}

\item{n}{number of estimation points for each 1-dimensional coefficient
function. If estimating more than one coefficient, can be entered
as a vector of length equal to the number of coefficients. Default
is the number of unique observation points.}

\item{n2}{Square root of the number of points used to grid estimates of
2-dimensinoal functions for contouring.
If estimating more than one coefficient, can be entered
as a vector of length equal to the number of coefficients.}

\item{exclude}{if \code{TRUE}, excludes reporting of the estimate at
coordinates that are "too far" from data used to fit the model, as
determined by \code{mgcv::plot.mgcv.smooth}. Excluding these values sets
them to \code{NA} in the return object.}

\item{limit}{if \code{TRUE}, checks if a \code{limits} function was
used to generate the term, and if so, applies the function to the
output to only produce estimates over regions that were within "limits".}

\item{plotMe}{if \code{TRUE}, also plots the coefficient(s)}

\item{...}{further arguments passed on to \pkg{mgcv}'s \code{plot.gam}; see
  Details.}
}
\value{
a list of data.frames containing the evaluation points,
   coefficient function values and optionally their se's for each term in \code{select}.
   If only one term is selected, the one data frame is unlisted and the return
   object will be the single data frame.
}
\description{
This function is used to extract coefficients from a fitted `pcox` model, in
particular smooth functions resulting from including smooth effects
of scalar covariates, time-varying effects, functional predictors, or
historical effects. The function calls \code{mgcv::plot.gam} to generate
the smooth functions.
}
\details{
The smooth coefficients are generated by \code{mgcv::plot.gam}. The
  following are common arguments of \code{plot.gam} that can be
  specified in the function call to \code{coefficients.pcox}:
  \describe{
    \item{select}{integer vector indicating the index of the desired smooth
    term in \code{object$pcox$smooth}. If \code{NULL} (the default), returns
    all smooth terms.}
    \item{se}{if \code{TRUE} (the default), returns pointwise standard error
    estimates. The covariance matrix to use (either \code{var} or
    \code{var2}) is specified by \code{use.var2}.}
    \item{n}{number of evaluation points for each 1-D plot. Only equally-spaced
    grids are possible.}
    \item{n2}{Square root of the number of points used to grid estimates of
    2-D functions.}
  }
  See \code{\link[mgcv]{plot.gam}} for a full list of possible arguments.
  Note that many of these have to do with plotting, and will thus not affect
  the resultant coefficient function.

  This function currently does not work to extract coefficient involving
  smooths over 3 or more variables. To extract these coefficients, you
  must do so manually, see \code{\link[mgcv]{PredictMat}}.
}
\author{
Fabian Scheipl and Jonathan Gellar, adapted from
  \code{refund::coefficients.pfr}
}

