% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coefficients.pcox.R
\name{coefficients.pcox}
\alias{coef.pcox}
\alias{coefficients.pcox}
\title{Extract coefficients from a fitted pcox model}
\usage{
\method{coefficients}{pcox}(object, raw = FALSE, term = NULL, n = NULL,
  n2 = NULL, inds = NULL, se = TRUE, limit = TRUE, untransform = TRUE,
  seWithMean = TRUE, ...)

\method{coef}{pcox}(object, raw = FALSE, term = NULL, n = NULL,
  n2 = NULL, inds = NULL, se = TRUE, limit = TRUE, untransform = TRUE,
  seWithMean = TRUE, ...)
}
\description{
Extract coefficients from a fitted pcox model
}

