% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coefficients.pcox.R
\name{coefficients.pcox}
\alias{coef.pcox}
\alias{coefficients.pcox}
\title{Extract coefficients from a fitted pcox model}
\usage{
\method{coefficients}{pcox}(object, raw = FALSE, term = NULL, n = NULL,
  n2 = NULL, se = TRUE, limit = TRUE, untransform = TRUE,
  seWithMean = TRUE, ...)

\method{coef}{pcox}(object, raw = FALSE, term = NULL, n = NULL,
  n2 = NULL, se = TRUE, limit = TRUE, untransform = TRUE,
  seWithMean = TRUE, ...)
}
\arguments{
\item{object}{a fitted \code{pfr} object as produced by \code{\link{pfr}()}.}

\item{raw}{if \code{TRUE}, returns the parameters used in the fitting of the
model. This would be the actual spline coefficients for smooth terms.}

\item{term}{integer to indicate the desired smooth term, according to its order
in \code{object$pcox$smooth}. A vector of term indices may be supplied.}

\item{n}{number of estimation points for each 1-dimensional coefficient
function.}

\item{n2}{Square root of the number of points used to grid estimates of
2-dimensinoal functions for contouring}

\item{se}{if \code{TRUE}, returns pointwise standard error estimates. Two
estimates of the pointwise standard error are returned, based on the two
estimates of the covariance matrix produced by \code{coxph}; see
\code{\link[survival]{coxph}} for details.}

\item{limit}{if \code{TRUE}, checks the \code{limits} function that was
used to create the term and applies it to filter out some of the
coordinates. Most relevant for variable-domain and historical functional
terms.}

\item{untransform}{if \code{TRUE},}
}
\description{
This method is used to extract coefficients from a fitted `pcox` model, in
particular functional coefficients resulting from including smooth effects
of scalar covariates, time-varying effects, functional predictors, or
historical effects.
}

