% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coefficients.pcox.R
\name{coefficients.pcox}
\alias{coef.pcox}
\alias{coefficients.pcox}
\title{Extract coefficient functions from a fitted pcox-object}
\usage{
\method{coefficients}{pcox}(object, select = 1, coords = NULL, n = NULL,
  se = ifelse(length(object$pcox$smooth) & select, TRUE, FALSE),
  seWithMean = FALSE, limit = TRUE, ...)

\method{coef}{pcox}(object, select = 1, coords = NULL, n = NULL,
  se = ifelse(length(object$pcox$smooth) & select, TRUE, FALSE),
  seWithMean = FALSE, limit = TRUE, ...)
}
\arguments{
\item{object}{return object from \code{\link{pcox}}}

\item{select}{integer indicating the index of the desired smooth term
in \code{object$pcox$smooth}. Enter 0 to request the raw coefficients
(i.e., \code{object$coefficients}) and standard errors (if \code{se==TRUE}).}

\item{coords}{named list indicating the desired coordinates where the
coefficient function is to be evaluated. Names must match the names of
\code{object$pcox$smooth[[selet]]$term}. If \code{NULL}, uses \code{n}
to generate equally-spaced coordinates.}

\item{n}{integer vector indicating the number of equally spaced coordinates
for each term. If length 1, the same number is used for each term.
Otherwise, the length must match \code{object$pcox$smooth[[select]]$dim}.}

\item{se}{if \code{TRUE}, returns pointwise standard error estimates. Defaults
to \code{FALSE} if raw coefficients are being returned; otherwise \code{TRUE}.}

\item{seWithMean}{if \code{TRUE} the standard errors include uncertainty about
the overall mean; if \code{FALSE}, they relate purely to the centered
smooth itself. Marra and Wood (2012) suggests that \code{TRUE} results in
better coverage performance for GAMs, though we have not tested this for
survival models.}

\item{limit}{if \code{TRUE}, checks if a \code{limits} function was
used to generate the term, and if so, applies the function to the
output to only produce estimates over regions that were within "limits".}

\item{...}{these arguments are ignored}
}
\value{
a data frame containing the evaluation points,
   coefficient function values and optionally the SE's for the term indicated
   by \code{select}.
}
\description{
This function is used to extract coefficients from a fitted `pcox` model, in
particular smooth functions resulting from including smooth effects
of scalar covariates, time-varying effects, functional predictors, or
historical effects.
}
\details{
The smooth coefficients are generated by \code{mgcv::plot.gam}. The
  following are common arguments of \code{plot.gam} that can be
  specified in the function call to \code{coefficients.pcox}:
  \describe{
    \item{select}{integer vector indicating the index of the desired smooth
    term in \code{object$pcox$smooth}. If \code{NULL} (the default), returns
    all smooth terms.}
    \item{se}{if \code{TRUE} (the default), returns pointwise standard error
    estimates. The covariance matrix to use (either \code{var} or
    \code{var2}) is specified by \code{use.var2}.}
    \item{n}{number of evaluation points for each 1-D plot. Only equally-spaced
    grids are possible.}
    \item{n2}{Square root of the number of points used to grid estimates of
    2-D functions.}
  }
  See \code{\link[mgcv]{plot.gam}} for a full list of possible arguments.
  Note that many of these have to do with plotting, and will thus not affect
  the resultant coefficient function.
}
\author{
Fabian Scheipl and Jonathan Gellar
}
\references{
Marra, G and S.N. Wood (2012) Coverage Properties of Confidence Intervals for
Generalized Additive Model Components. Scandinavian Journal of Statistics.
}

