% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coefficients.pcox.R
\name{coefficients.pcox}
\alias{coef.pcox}
\alias{coefficients.pcox}
\title{Extract coefficients from a fitted pcox model}
\usage{
\method{coefficients}{pcox}(object, raw = FALSE, term = NULL, n = NULL,
  n2 = NULL, se = TRUE, seWithMean = TRUE, exclude = FALSE,
  limit = TRUE, ...)

\method{coef}{pcox}(object, raw = FALSE, term = NULL, n = NULL,
  n2 = NULL, se = TRUE, seWithMean = TRUE, exclude = FALSE,
  limit = TRUE, ...)
}
\arguments{
\item{object}{a fitted \code{pcox} object as produced by \code{\link{pcox}()}.}

\item{raw}{if \code{TRUE}, returns the parameters used in the fitting of the
model. This would be the actual spline coefficients for smooth terms.}

\item{term}{integer to indicate the desired smooth term, according to its order
in \code{object$pcox$smooth}. A vector of term indices may be supplied.}

\item{n}{number of estimation points for each 1-dimensional coefficient
function.}

\item{n2}{Square root of the number of points used to grid estimates of
2-dimensinoal functions for contouring}

\item{se}{if \code{TRUE}, returns pointwise standard error estimates. Two
estimates of the pointwise standard error are returned, based on the two
estimates of the covariance matrix produced by \code{coxph}; see
\code{\link[survival]{coxph}} for details.}

\item{seWithMean}{if \code{TRUE} the component smooths are shown with
confidence intervals that include the uncertainty about the overall mean;
if \code{FALSE}, then the uncertainty relates purely to the centered
smooth itself.}

\item{exclude}{if \code{TRUE}, excludes reporting of the estimate at coordinates that are
"too far" from data used to fit the model, as determined by
\code{mgcv::plot.mgcv.smooth}, by setting the estimate to \code{NA}.}

\item{limit}{if \code{TRUE}, checks the \code{limits} function that was
used to create the term and applies it to filter out some of the
coordinates. Most relevant for variable-domain and historical functional
terms.}

\item{...}{other parameters to pass to the plotting function (either
  \code{mgcv:::plot.mgcv.smooth} or \code{mgcv:::plot.random.effect}.}
}
\value{
a data frame containing:
  \itemize{
    \item The coordinates where the coefficient is estimated
    \item The estimated coefficient at those coordinates (\code{value})
    \item The standard error estimate(s) at those coordinates, if requested
  }
}
\description{
This method is used to extract coefficients from a fitted `pcox` model, in
particular smooth functions resulting from including smooth effects
of scalar covariates, time-varying effects, functional predictors, or
historical effects.
}
\details{
This function calls either \code{mgcv:::plot.mgcv.smooth} or
  \code{mgcv:::plot.random.effect} to get the data that would be used
  to create an \code{mgcv}-style plot, but instead of plotting the data,
  it returns it. Much of the code and ideas behind this method were taken
  from \code{refund::coefficients.pfr}.
}
\author{
Jonathan Gellar <jgellar1@jhu.edu> and Fabian Scheipl
}
\seealso{
\code{\link[mgcv]{plot.gam}}, \code{refund::coefficients.pfr}
}

