% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.tt.func}
\alias{create.tt.func}
\title{Create tt function for a time-varying terms}
\usage{
create.tt.func(limits, linear, tv, basistype, sind, integration, standardize,
  s.transform, t.transform, basisargs, method, eps, map)
}
\description{
Create tt function for a time-varying terms
}
\keyword{internal}

