% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.tt.func.R
\name{create.tt.func}
\alias{create.tt.func}
\title{Create \code{tt} function for a time-varying term}
\usage{
create.tt.func(limits, linear, tv, basistype, sind, integration, standardize,
  s.transform, t.transform, basisargs, method, eps, map)
}
\arguments{
\item{limits}{specifies the term as either a term involving scalar
covariates, a concurrent effect of a time-varying covariate, a baseline
functional covariate, or a historical effect of a time-varying covariate.
Defaults to \code{NULL}, indicating scalar covariates. See Details.}

\item{linear}{if \code{FALSE}, covariates are included as nonlinear (smooth)
effects, otherwise as a linear effect.}

\item{tv}{if \code{TRUE}, makes the effect time-varying.}

\item{basistype}{specifies the basis constructor function (from the
\code{mgcv} package) that is used to define a smooth term. Defaults to
\code{\link[mgcv]{s}}, which is the only option allowed for smooths of
only one argument. For smooths of multiple arguments (including t and s),
\code{\link[mgcv]{te}} or \code{\link[mgcv]{t2}} may (but don't have to)
be used.}

\item{sind}{specifies the time indices for functional and time-varying
predictors. Can be entered as a vector of length \code{ncol(X)}, or a
matrix of the same dimensions as \code{X} (for covariates measured on
unequal grids).}

\item{integration}{method for numerical integration.}

\item{standardize}{if \code{TRUE}, the term is "standardized" by dividing
by the width of integration.}

\item{s.transform}{optional transformation function for the first variable
of the smooth. For functional/historical predictors, this is the variable
over which the integration takes place.}

\item{t.transform}{optional transformation function for the time variable,
if it is one of the indices of the smooth}

\item{basisargs}{arguments for the function specified by \code{basistype},
used to set up the basis and penalization}

\item{method}{method of optimization of the smoothing parameter, for
penalized terms}

\item{eps}{tolerance level for the optimization of the smoothing parameter,
for penalized terms}

\item{map}{variable map used to turn the \code{x} argument of the \code{tt}
  function from a matrix back into a data frame. Necessary because \code{x}
  must be a matrix, but we need to know which columns correspond to which
  variables.}
}
\value{
a \code{tt} function.
The \code{tt} function takes two arguments, \code{x} (the data) and \code{t}
(the time vector). For terms involving time-varying effects or covariates,
these arguments will already be expanded (by \code{coxph}) to "long" format
at the time the \code{tt} function is called. According to what is entered
for the \code{limits} argument, the \code{tt} function will do one of the
following:
\enumerate{
  \item For terms that do not involve smooths, it simply returns the model
        matrix columns for this term
  \item For terms that do involve smooths, it calls \code{pcoxTerm} and
        returns the result.
}

Note that the \code{tt} function will contain all the information needed
to process the data in its environment. For this reason, \code{pcox} saves
the \code{tt} function as part of its return object, so that any methods
called on that object (e.g., \code{predict} or \code{coef}) know how to
process the data.
}
\description{
A time-transform (\code{tt}) function is used by \code{survival::coxph} to
allow variables to vary dynamically in time. This allows both time-varying
covariates and time-varying effects to be included in a \code{coxph} (and
thus \code{pcox}) formula. This function sets up the \code{tt} function
according to the supplied arguments, and returns that function for later
use by \code{coxph}.
}
\seealso{
\code{\link[survival]{coxph}}, \code{\link{p}},
  \code{\link{pcoxTerm}}, \code{\link{create.xt.func}}
}
\keyword{internal}

