% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.tt.p.R
\name{create.tt.p}
\alias{create.tt.p}
\title{Create tt function for a time-varying terms}
\usage{
create.tt.p(limits, linear, tv, basistype, sind, integration, standardize,
  domain, basisargs, method, eps, map)
}
\description{
Create tt function for a time-varying terms
}
\keyword{internal}

