% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.xt.func.R
\name{create.xt.func}
\alias{create.xt.func}
\title{Create x-transform function for non-time-varying terms}
\usage{
create.xt.func(limits, linear, basistype, sind, integration, standardize,
  basisargs, method, eps)
}
\arguments{
\item{limits}{specifies the term as either a term involving scalar
covariates or a functional covariate.
Defaults to \code{NULL}, indicating scalar covariates. See Details.}

\item{linear}{if \code{FALSE}, covariates are included as nonlinear (smooth)
effects, otherwise as a linear effect.}

\item{basistype}{specifies the basis constructor function (from the
\code{mgcv} package) that is used to define a smooth term. Defaults to
\code{\link[mgcv]{s}}, which is the only option allowed for smooths of
only one argument. For smooths of multiple arguments,
\code{\link[mgcv]{te}} or \code{\link[mgcv]{t2}} may (but don't have to)
be used.}

\item{sind}{specifies the argument values for functional predictors. Can be
entered as a vector of length \code{ncol(X)}, or a matrix of the same
dimensions as \code{X} (for covariates measured on unequal grids).}

\item{integration}{method for numerical integration.}

\item{standardize}{if \code{TRUE}, the term is "standardized" by dividing
by the width of integration.}

\item{basisargs}{arguments for the function specified by \code{basistype},
used to set up the basis and penalization}

\item{method}{method of optimization of the smoothing parameter, for
penalized terms}

\item{eps}{tolerance level for the optimization of the smoothing parameter,
  for penalized terms}
}
\value{
An \code{xt} function.
The \code{xt} function takes one argument, \code{x} (the data). According
to what is entered for the \code{limits} argument, the \code{xt} function
will return one of the following:
\enumerate{
  \item For terms that do not involve smooths, it simply returns the model
        matrix columns for this term
  \item For terms that do involve smooths, it calls \code{pcoxTerm} and
        returns the result.
}

Note that the \code{xt} function will contain all the information needed
to process the data in its environment. For this reason, \code{pcox} saves
the \code{xt} function as part of its return object, so that any methods
called on that object (e.g., \code{predict} or \code{coef}) know how to
process the data.
}
\description{
The \code{xt} function processes data that does not involve time
transformations for inclusion in the \code{coxph} formula. Unlike a \code{tt}
function for time-transformed terms (see \code{\link{create.tt.func}}),
this function is not required by \code{coxph}. However, we use it for
consistency with how we handle \code{tt} terms, and also because it provides
a convenient place to store all information needed to process the data
and set up any required basis.
}
\seealso{
\code{\link[survival]{coxph}}, \code{\link{p}},
  \code{\link{pcoxTerm}}, \code{\link{create.tt.func}}
}
\keyword{internal}

