% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.xt.func.R
\name{create.xt.func}
\alias{create.xt.func}
\title{Create x-transform function for non-time-varying terms}
\usage{
create.xt.func(limits, linear, basistype, sind, integration, standardize,
  s.transform, t.transform, basisargs, method, eps)
}
\description{
Create x-transform function for non-time-varying terms
}
\keyword{internal}

