% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcox-Accessory.R
\name{f_override}
\alias{f_override}
\title{Accessory function that allows s(), te(), and t2() in coxph term names}
\usage{
f_override(...)
}
\description{
Returns the variable that matches the call, instead of calling the
corresponding function. Useful for allowing special term names.
}
\keyword{internal}

