% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/f_override.R
\name{f_override}
\alias{f_override}
\title{Accessory function that allows s(), te(), and t2() in coxph term names}
\usage{
f_override(...)
}
\description{
Accessory function that allows s(), te(), and t2() in coxph term names
}
\keyword{internal}

