% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getROC.R
\name{getROC}
\alias{getROC}
\title{Generate Time-Varying ROC Curve}
\usage{
getROC(fit, utime = NULL, evaltimes = NULL, plot = FALSE, K = NULL,
  data = NULL, bhaz = c("none", "breslow"), St = NULL, tmax = NULL)
}
\description{
Generate Time-Varying ROC Curve
}
\keyword{internal}

