% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeBetaMat.R
\name{makeBetaMat}
\alias{makeBetaMat}
\title{Functions for simulating beta coefficient matrices}
\usage{
makeBetaMat(J = 101, bfunc = genBeta1, ...)
}
\arguments{
\item{J}{Width of function}

\item{bfunc}{the function defining the bivariate coefficient function}

\item{...}{other arguments to pass to \code{bfunc}}
}
\description{
Functions for simulating beta coefficient matrices
}

