% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pcox}
\alias{pcox}
\title{Fit a Penalized Cox Regression Model}
\usage{
pcox(formula, method = c("aic", "caic", "epic", "reml", "ml", "fixed", "df"),
  eps = 0.001, knots = NULL, ...)
}
\arguments{
\item{formula}{a formula with special terms as for \code{\link[mgcv]{gam}}, with
additional special terms allowed. See Details.}

\item{method}{method used to optimize the smoothing parameter(s). See Details.}

\item{eps}{convergence level for the criterion indicated by \code{method}.}

\item{knots}{optional list containing the user-specified knot values for basis
construction, as in \code{\link[mgcv]{gam}}.}
}
\value{
a fitted \code{pcox} object. This is either a \code{coxph} or \code{coxme} object with
   additional information in the \code{pcox} entry.
}
\description{
This function fits a penalized Cox regression model using penalized splines
to model the smooth coefficients. It is a wrapper
for \code{survival::coxph} and \code{coxme::coxme} to fit Cox models with
smooth covariate effects for (possibly censored) time-to-event data, of
the general form
\eqn{log[h_i(t)] = log[h_0(t)] + f_1(z_i, t) + f_2(X_i(s), t)}
where \eqn{h_i(t)} is the hazard function for subject i, \eqn{h_0(t)} is
the baseline hazard function, \eqn{z_i} are scalar covariates, and
\eqn{X_i(s)} are functional covariates.
}
\details{
This routine is a wrapper for either \code{\link[survival]{coxph}} or
\code{\link[coxme]{coxme}}, which fit penalized survival models by maximizing
the penalized partial likelihood. The fitting routine is dependent on the
\code{method} chosen for optimizing the smoothing parameter: ML or REML optimization
is performing by \code{coxme}, whereas likelihood-based criteria (AIC, AIC_c, or EPIC)
are implemented by \code{coxph}. \code{coxph} is also used to implement the penalized
survival model with fixed value of the smoothing parameter (\code{method}="fixed", and
a \code{theta} arguemnt is required) or degrees of freedom (\code{method}="df",
and a \code{df} argument is required).

Smooth (penalized) terms are indicated in the model formula as follows:
\enumerate{
   \item Nonlinear (and possibly multivariate) effects of scalar covariates that do
     not vary with the time index \eqn{t} (i.e., f(z_{i1})) are specified using either
     \code{\link{s}}(), or \code{\link{te}}(), as in \code{\link[mgcv]{gam}}.
   \item Nonlinear (and possibly multivariate) effects of scalar covariates that vary
     with the time index \eqn{t} (i.e., f(z_{i1},t)) are specified using ???.
   \item Linear and nonlinear effects of non-concurrent functional predictors that
     do not vary with the time index \eqn{t} (i.e., \eqn{\int X_i(s)\beta(s)ds} or
     \eqn{\int\beta(X_i(s), s) ds}) are specified using \code{\link[refund]{lf}},
     \code{\link[refund]{af}}, or \code{\link[refund]{lf.vd}} from the package
     \code{refund}.
   \item Linear and nonlinear effects of non-concurrent functional predictors that vary
     with the time index \eqn{t} are specified using \code{\link[refund]{ff}} or
     \code{\link[refund]{sff}} from the package \code{refund}, respectively. Note that
     the outcome "function" for these "function-on-function" terms is the hazard function.
   \item Concurrent effects of functional covariates are not yet allowed, but will be soon!
}

Note that we refer to a functional predictor as "concurrent" if it is measured over the
domain \eqn{t}, the same domain as the outcome event is measured across. Another
term for this type of predictor is a densely-measured time-varying covariate. These
types of covariates require special attention....
}
\author{
Jonathan Gellar <jgellar1@jhu.edu>
}
\references{
Gellar, Jonathan E., Colantuoni, Elizabeth, Needham, Dale M., and Crainiceanu,
   Ciprian M. (2014). "Cox Regression Models with Functional Covariates for Survival Data."
   Johns Hopkins University, Dept. of Biostatistics Working Papers. Working Paper 264.
   \link{http://biostats.bepress.com/jhubiostat/paper264}.
}
\seealso{
\code{mgcv}'s \code{\link[mgcv]{smooth.terms}} for details of \code{mgcv} syntax and
   available spline bases and penalties; the related \code{\link[refund]{pffr}} and
   \code{\link[refund]{fgam}} from \code{refund}.
}

