% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcox-package.R
\docType{package}
\name{pcoxp}
\alias{pcox-package}
\alias{pcoxp}
\alias{pcoxp-package}
\title{Penalized Cox regression models}
\description{
\code{\link{pcox}()} fits penalized Cox regression models using penalized splines
to model the smooth coefficients.
}

