% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pcoxTerm.R
\name{pcoxTerm}
\alias{pcoxTerm}
\title{Create a smooth and coxph.penalty object for use in a pcox formula}
\usage{
pcoxTerm(data, limits, linear, tv, basistype, sind, integration, standardize,
  s.transform, t.transform, basisargs, method, eps, smooth, s0, t0 = NULL,
  t = NULL)
}
\description{
Create a smooth and coxph.penalty object for use in a pcox formula
}

