% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcox-Accessory.R
\name{safeDeparse}
\alias{safeDeparse}
\title{Deparse a function safely}
\usage{
safeDeparse(expr)
}
\description{
This utility function was copied from the \code{refund} package.
}
\keyword{internal}

